+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Overriding `matchmaker.docker.image_id` getting ignored
 - `rivet config validate` now uses `--print` flag instead of a positional argument
+- Only delete `.rivet/config.*` to keep other files in `./.rivet` directory
 
 ## [v1.0.0-rc.1] - 2023-12-24